move get/delete commands to files namespace, make get take the output as argument instead of option
