+++ apps/files/appinfo/info.xml
@@ -35,6 +35,8 @@
 		<command>OCA\Files\Command\TransferOwnership</command>
 		<command>OCA\Files\Command\ScanAppData</command>
 		<command>OCA\Files\Command\RepairTree</command>
+		<command>OCA\Files\Command\Get</command>
+		<command>OCA\Files\Command\Delete</command>
 	</commands>
 
 	<activity>
